Fixed getting previous file when running
My mistake in the first place was putting the creation of the file inside the MonoCompiler class where the exe file is executed. So what I needed to do is that I needed to move the creation of the exe file inside the UpdateFiles class so that the exe file is created every time the code is saved.