implemented a starting scene cleanup
this is needed because when the client or the host disconnects and goes back to the starting scene the networkManager object continues to exist and when we try to join or create a new game it won't let us because it will try to create a new one and this is not allowed. There can only exist one networkMangerObject.